qwiz-app/backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests
pydantic
pydantic-settings

# Google Cloud
firebase-admin
google-cloud-firestore

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
pytest-mock==3.12.0